requirements.txt
python-telegram-bot[job-queue]==22.6
anthropic>=0.40.0
openai>=1.50.0
python-dotenv>=1.0.0
